simplified action return values

+++ lib/Threads/Action/DeleteSubscriptions.pm
@@ -3,9 +3,8 @@ package Threads::Action::DeleteSubscriptions;
 use strict;
 use warnings;
 
-use parent 'Tu::Action';
+use parent 'Threads::Action';
 
-use Threads::DB::User;
 use Threads::DB::Subscription;
 
 sub run {
@@ -13,14 +12,11 @@ sub run {
 
     my $user = $self->scope->user;
 
-    Threads::DB::Subscription->table->delete(
-        where => [
-            user_id => $user->id,
-        ]
-    );
+    Threads::DB::Subscription->table->delete(where => [user_id => $user->id]);
 
-    return {redirect => $self->url_for('list_subscriptions') . ''},
-      type => 'json';
+    my $url = $self->url_for('list_subscriptions');
+
+    return $self->new_json_response(200, {redirect => "$url"});
 }
 
 1;
